+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Added the `@click` event to select the input text in `QasNumericInput`.
+
+### Changed
+- Changed `@input` to emit the `update-model` event in `QasNumericInput`. The event arguments contains the raw and the unformatted value.
+- Changed the behavior of the `QasNumericInput` to emit the updated value in `onBlur` event.
+
 ## 2.19.0 - 2022-05-05
 
 ### Added